Sun, Aug 11, 2024, 11:43 AM +02:00

+++ readability/readability.c
@@ -21,7 +21,7 @@ int main(void)
     // Compute the Coleman-Liau index
     float l = ((float) letters / words) * 100;
     float s = ((float) sentences / words) * 100;
-    int index = 0.0588 * l - 0.296 * s - 15.8;
+    float index = 0.0588 * l - 0.296 * s - 15.8;
     // Print the grade level
     printf("%d\n", letters);
     printf("%.2f\n", index);
